Use immutable textures for render textures with Emscripten
This is a workaround for this issue with Chromium: https://issues.chromium.org/issues/355605685
Effects that used offscreen rendering stopped working. Even thought this is fixed in
new version of Chrome, this is a big issue for Safari as well which uses that version of ANGLE
library that has this bug on iOS 18.
Immutable textures are created with glTexStorage2D and filled with glTexSubImage2D.

Signed-off-by: Hrvoje Ditrih <hrvoje.ditrih@gmail.com>